requirements.txt
django-extensions==2.2.1